term: add DECSCA character protection + VT420 rectangular area ops (#71)
DECSCA (CSI Ps " q) marks characters protected. Only the selective
erases honor it: DECSEL (CSI ? Ps K), DECSED (CSI ? Ps J), DECSERA
(CSI ... $ {). All other erases, scroll, and overwrites ignore it.

New VT420 rectangular operations:
- DECERA (CSI ... $ z) — erase rectangular area
- DECFRA (CSI ... $ x) — fill rectangular area
- DECSERA (CSI ... $ {) — selective erase rectangular area
- DECCARA (CSI ... $ r) — change attributes in rectangular area
- DECRARA (CSI ... $ t) — reverse attributes in rectangular area
- DECCRA (CSI ... $ v) — copy rectangular area
- DECSACE (CSI Ps * x) — attribute change extent (stream/rectangle)
- DECRQSS answers for "q (DECSCA) and *x (DECSACE)

Protection lives in CurAttrs bit attrProtected — SGR 0 does not clear
it; DECSC/DECRC and alt-screen swap carry it; DECSTR and RIS reset it.
Cell Attrs widened from uint8 to uint16 for the extra bit; attrVisual
mask isolates render-relevant bits so blank-cell fast paths treat a
protected space as untouched.

+++ term/grid_cursor.go
@@ -27,12 +27,13 @@ func (g *grid) ApplyDECSCUSR(ps int) {
 }
 
 // savedCursor holds the snapshot taken by SaveCursor (DECSC / CSI s).
-// Stores position and SGR state per VT100 spec. Zero value means no
-// snapshot has been taken yet (valid == false).
+// Stores position and SGR state per VT100 spec. attrs carries the DECSCA
+// protection bit along with the SGR attributes, which VT510 requires DECSC
+// to save. Zero value means no snapshot has been taken yet (valid == false).
 type savedCursor struct {
 	r, c       int
 	fg, bg     uint32
-	attrs      uint8
+	attrs      uint16
 	ulStyle    uint8
 	ulColor    uint32
 	charsetG0  byte
